requirements.txt
# Core dependencies
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1

# State management
# Redis 5.0+ includes cluster support and async functionality built-in
redis==5.0.1

# Kubernetes client
kubernetes==28.1.0

# Monitoring & metrics
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0

# Data processing
numpy==1.26.2
pillow==10.1.0
opencv-python==4.9.0.80

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-benchmark==4.0.0
pytest-cov==4.1.0
locust==2.20.0

# Utilities
pydantic==2.5.2
python-json-logger==2.0.7
tenacity==8.2.3
